requirements.txt
# Smart Agriculture AI - Dependencies
# Python 3.8+ recommended

# Web Framework
streamlit==1.29.0

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Weather API
requests>=2.31.0

# Database
# SQLite is built-in to Python

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
